requirements.txt
streamlit
pymupdf
scikit-learn
nltk
pandas
pytesseract
Pillow
